+++ lib/main.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:grupp_5/components/models/recipe_model.dart';
-import 'package:grupp_5/components/models/steps_model.dart';
+import 'package:grupp_5/components/providers/provider.dart';
 import 'package:grupp_5/views/dev_view.dart';
 import 'package:grupp_5/views/save_view.dart';
 import 'package:provider/provider.dart';
 import '/constants/routes.dart';
 import 'views/filter_view.dart';
-
 import 'views/recipe_view.dart';
 import 'views/scramble_view.dart';
 import 'views/info_view.dart';
@@ -17,8 +15,6 @@ void main() {
     MultiProvider(
       providers: [
         ChangeNotifierProvider(create: (context) => RecipeProvider()),
-        ChangeNotifierProvider(
-            create: (context) => AnalyzedInstructionProvider()),
       ],
       child: MaterialApp(
         theme: ThemeData(
