+++ lib/views/scramble_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:grupp_5/components/models/api_service.dart';
 import 'package:grupp_5/components/models/recipe_model.dart';
 import 'package:grupp_5/components/models/steps_model.dart';
+import 'package:grupp_5/components/providers/provider.dart';
 import 'package:provider/provider.dart';
 import '../constants/constants.dart';
 import '/constants/routes.dart';
@@ -193,6 +194,7 @@ class _ScrambleViewState extends State<ScrambleView> {
               setState(() {
                 apiId = Random().nextInt(5000);
                 recipe.fetchRecipe();
+                recipe.fetchAnalyzedInstruction();
               });
             },
             child: Container(
